+++ libs/opengl/include/mrpt/opengl/CCamera.h
@@ -2,7 +2,7 @@
    |                     Mobile Robot Programming Toolkit (MRPT)            |
    |                          https://www.mrpt.org/                         |
    |                                                                        |
-   | Copyright (c) 2005-2021, Individual contributors, see AUTHORS file     |
+   | Copyright (c) 2005-2022, Individual contributors, see AUTHORS file     |
    | See: https://www.mrpt.org/Authors - All rights reserved.               |
    | Released under BSD License. See: https://www.mrpt.org/License          |
    +------------------------------------------------------------------------+ */
@@ -107,7 +107,7 @@ class CCamera : public CRenderizable
 		m_pointingY = p.y();
 		m_pointingZ = p.is3DPoseOrPoint() ? p.m_coords[2] : 0;
 	}
-	inline void setPointingAt(const mrpt::math::TPoint3D& p)
+	void setPointingAt(const mrpt::math::TPoint3D& p)
 	{
 		setPointingAt(d2f(p.x), d2f(p.y), d2f(p.z));
 	}